DeepSARSA/requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.5
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
flatbuffers==22.11.23
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
importlib-resources==5.2.0
ipykernel==6.15.2
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.4.7
jupyter-console==6.4.3
jupyter_core==4.11.2
jupyter-server==1.18.1
jupyterlab==3.5.0
jupyterlab-pygments==0.1.2
jupyterlab_server==2.16.3
jupyterlab-widgets==1.0.0
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
ply==3.11
priority==2.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.19.6
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
pygbag==0.6.0
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.0
quart==0.18.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
Send2Trash==1.8.0
setuptools==65.5.0
sip==6.6.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.13.1
tifffile==2022.10.10
tinycss2==1.2.1
tk==0.1.0
token-utils==0.1.8
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.1.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.8.0
